+++ packages/dialtone-vue2/components/popover/popover.vue
@@ -528,6 +528,7 @@ export default {
       POPOVER_PADDING_CLASSES,
       POPOVER_HEADER_FOOTER_PADDING_CLASSES,
       intersectionObserver: null,
+      mutationObserver: null,
       isOutsideViewport: false,
       isOpen: false,
       anchorEl: null,
@@ -627,10 +628,10 @@ export default {
     isOpen (isOpen, isPrev) {
       if (isOpen) {
         this.initTippyInstance();
-        this.tip.show();
+        this.tip?.show();
       } else if (!isOpen && isPrev !== isOpen) {
         this.removeEventListeners();
-        this.tip.hide();
+        this.tip?.hide();
       }
     },
   },
@@ -638,16 +639,11 @@ export default {
   mounted () {
     warnIfUnmounted(this.$el, this.$options.name);
 
-    const externalAnchorEl = this.externalAnchor
-      ? this.$refs.anchor.getRootNode().querySelector(`#${this.externalAnchor}`)
-      : null;
-    this.anchorEl = externalAnchorEl ?? this.$refs.anchor.children[0];
     this.popoverContentEl = this.$refs.content?.$el;
+    this.updateAnchorEl();
 
-    if (this.isOpen) {
-      this.initTippyInstance();
-      this.tip.show();
-    }
+    this.mutationObserver = new MutationObserver(this.updateAnchorEl);
+    this.mutationObserver.observe(this.$refs.anchor, {childList: true});
 
     // rootMargin here must be greater than the margin of the height we are setting in calculatedMaxHeight which
     // currently is var(--dt-space-300) (4px). If not the intersectionObserver will continually trigger in an infinite
@@ -660,6 +656,7 @@ export default {
   beforeDestroy () {
     this.tip?.destroy();
     this.intersectionObserver?.disconnect();
+    this.mutationObserver?.disconnect();
     this.removeReferences();
     this.removeEventListeners();
   },
@@ -673,6 +670,26 @@ export default {
       if (!dialog) return;
       const isOut = isOutOfViewPort(dialog);
       this.isOutsideViewport = isOut.bottom || isOut.top;
+    },
+
+    updateAnchorEl () {
+      const externalAnchorEl = this.externalAnchor
+        ? this.$refs.anchor.getRootNode().querySelector(`#${this.externalAnchor}`)
+        : null;
+      this.anchorEl = externalAnchorEl ?? this.$refs.anchor.children[0];
+
+      this.tip?.destroy();
+      delete this.tip;
+
+      if (!this.anchorEl) {
+        console.warn('No anchor found for popover');
+        return;
+      }
+
+      if (this.isOpen) {
+        this.initTippyInstance();
+        this.tip?.show();
+      }
     },
 
     popperOptions () {
@@ -718,7 +735,7 @@ export default {
 
       this.isOpen = true;
       await this.$nextTick();
-      this.tip.setProps({
+      this.tip?.setProps({
         placement: 'right-start',
         getReferenceClientRect: () => ({
           width: 0,
@@ -776,7 +793,7 @@ export default {
         if (!element) return;
         if (element.tagName?.toLowerCase() === 'body') {
           disableRootScrolling(this.anchorEl.getRootNode().host);
-          this.tip.setProps({ offset: this.offset });
+          this.tip?.setProps({ offset: this.offset });
         } else {
           element.classList.add('d-zi-popover');
         }
@@ -791,7 +808,7 @@ export default {
       if (!element) return;
       if (element.tagName?.toLowerCase() === 'body') {
         enableRootScrolling(this.anchorEl.getRootNode().host);
-        this.tip.setProps({ offset: this.offset });
+        this.tip?.setProps({ offset: this.offset });
       } else {
         element.classList.remove('d-zi-popover');
       }
@@ -960,6 +977,8 @@ export default {
           internalAppendTo = this.appendTo;
           break;
       }
+
+      this.tip?.destroy();
 
       this.tip = createTippyPopover(this.anchorEl, {
         popperOptions: this.popperOptions(),