+++ packages/dialtone-vue3/components/popover/tippy_utils.js
@@ -66,7 +66,11 @@ const createAnchor = (anchorWrapper) => {
 };
 
 export const getAnchor = (anchorWrapper) => {
-  const anchor = anchorWrapper?.children[0];
+  if (!anchorWrapper) {
+    console.warn('No anchor wrapper provided. This may cause issues with the popover.');
+    return;
+  }
+  const anchor = anchorWrapper.children[0];
   if (!anchor) return createAnchor(anchorWrapper);
   return anchor;
 };